+++ src/api/routes/tokens/tokens.ts
@@ -22,10 +22,7 @@ import { bufferToHexPrefixString, has0xPrefix, isValidPrincipal } from '../../..
 import { booleanValueForParam, isUnanchoredRequest } from '../../../api/query-helpers';
 import { decodeClarityValueToRepr } from 'stacks-encoding-native-js';
 import { getAssetEventTypeString, parseDbTx } from '../../controllers/db-controller';
-import {
-  getChainTipCacheHandler,
-  setChainTipCacheHeaders,
-} from '../../controllers/cache-controller';
+import { getETagCacheHandler, setETagCacheHeaders } from '../../controllers/cache-controller';
 
 const MAX_TOKENS_PER_REQUEST = 200;
 const parseTokenQueryLimit = parseLimitQuery({
@@ -35,7 +32,7 @@ const parseTokenQueryLimit = parseLimitQuery({
 
 export function createTokenRouter(db: DataStore): express.Router {
   const router = express.Router();
-  const cacheHandler = getChainTipCacheHandler(db);
+  const cacheHandler = getETagCacheHandler(db);
   router.use(express.json());
 
   router.get(
@@ -98,7 +95,7 @@ export function createTokenRouter(db: DataStore): express.Router {
         total: total,
         results: parsedResults,
       };
-      setChainTipCacheHeaders(res);
+      setETagCacheHeaders(res);
       res.status(200).json(response);
     })
   );
@@ -159,7 +156,7 @@ export function createTokenRouter(db: DataStore): express.Router {
         total: total,
         results: parsedResults,
       };
-      setChainTipCacheHeaders(res);
+      setETagCacheHeaders(res);
       res.status(200).json(response);
     })
   );
@@ -214,7 +211,7 @@ export function createTokenRouter(db: DataStore): express.Router {
         total: total,
         results: parsedResults,
       };
-      setChainTipCacheHeaders(res);
+      setETagCacheHeaders(res);
       res.status(200).json(response);
     })
   );